requirements.txt
opencv-python
pyserial
numpy
aiortc
google-genai
aiohttp
aiohttp-cors
pyusbcameraindex
